First Round Write Up
This is a brief of the finding for the first round challenge, where we worked on predicting a Lyft's driver's average lifetime value and offering impactful suggestions on Lyft's business based on our findings from the data. 
Selected out of 300+ teams and became one of the final 16 teams invited to present and compete onsite in Lyft's Headquarter on October, 14th.